src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/remove-styles-in-shadow-crash-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: host.createShadowRoot is not a function
PASS
